refactor: create ostree repo in ll-package-manager
Now we create /var/lib/linglong and chown it to linglong user in
systemd service file. We add a script call create-linglong-dirs to do
this job.

As script might be installed to /usr/lib/linglong or
/usr/libexec/linglong. I just ignore the error.

Service files should be configured in cmake, let's fix this later.

OstreeRepoHelper has method ensureRepoEnv which will create the ostree
repo if it doesn't exist. This should be fine.

Update1: But we indeed have some codes open the ostree repo before call
OSTreeRepoHelper's ensureRepoEnv method. This commit include some quick
hack to fix this issue, which should be refactor later.

Update2: ensureRepoEnv use ostree_repo_new which does not create the
ostree repo if it not exist. This behavior is not match comment. So I
refactor this method to make it do the job it should have done.