fix: install app with default verison is wrong
If the app is already installed, we should compare the local
version with the remote version. And the remote version of app
should not be installed if the local version is bigger than the
remote version.

Log:

(cherry picked from commit e9dd54d07561c515bb1393ce1bb48f48b27fa618)