fix(cmake): install linglong-repair-tool as a program
Previously, misc/CMakeLists.txt installed the program linglong-repair-tool as
a "FILES" item (no executable bit), this is incorrect as the user would fail
to run this program (with a "permission denied" error).

Correct the INSTALL() command and install linglong-repair-tool as a "PROGRAMS"
item to give it executable bit at install-time.