fix(box): default to follow symbol link when parse mount source
When we lookup the source of a mount, unless we get a mount option tell
us to not follow symbol link, which means to treat the link it self as
a regular file, we should always follow the link.

For example, if the /etc in runtime namespace is a symbol link to
/usr/etc, and config.json tell us to perform a bind mount from runtime
namespace /etc to container namespace /etc, then the old code will
create /etc in container namespace as a file, which leads to an ENOTDIR
error when doing the bind mount.

Bug: https://pms.uniontech.com/bug-view-274969.html