fix: print invalid task when the task is queued
The initial State in the cli is Unknown, and the initial
State of PackageTask is Queued. When a new task enters and
queues, PM will update the State of task to Queued, but the
State has not actually changed at this time.
After receiving the PropertyChanged signal, and the cli will
not update the states from Unknown to Queued. Then the cli finds
that its State is Unknown, so it prints invalid task.

Log